+++ src/chromyRunner.js
@@ -5,12 +5,10 @@ import actions from './actions';
 import { configTypes } from './defaultConfig';
 
 const saveImage = (filename, image, testType, screenshotsPath, testReportPath) => {
-  if (testType === configTypes.test) {
-    logger.log(`screenshot saved in -> ${testReportPath}/${filename}.png`);
-    return fs.writeFileSync(`${testReportPath}/${filename}.png`, image);
-  }
-  logger.log(`screenshot saved in -> ${screenshotsPath}/${filename}.png`);
-  return fs.writeFileSync(`${screenshotsPath}/${filename}.png`, image);
+  const directory = testType === configTypes.test ? testReportPath : screenshotsPath;
+  const filePath = `${directory}/${filename}.png`;
+  logger.log(`screenshot saved in -> ${filePath}`);
+  return fs.writeFileSync(filePath, image);
 };
 
 const run = async (chromy, options, test) => {
@@ -32,7 +30,7 @@ const run = async (chromy, options, test) => {
             try {
               logger.log('Capturing screenshot of whole DOM');
               const png = await chromy.screenshotDocument();
-              await saveImage(test.name, png, test.type, options.screenshots, options.testReportPath);
+              saveImage(test.name, png, test.type, options.screenshots, options.testReportPath);
             } catch (error) {
               logger.error(error);
               return false;
@@ -42,7 +40,7 @@ const run = async (chromy, options, test) => {
             try {
               logger.log('Capturing screenshot of chrome window');
               const png = await chromy.screenshot();
-              await saveImage(test.name, png, test.type, options.screenshots, options.testReportPath);
+              saveImage(test.name, png, test.type, options.screenshots, options.testReportPath);
             } catch (error) {
               logger.error(error);
               return false;
@@ -52,7 +50,7 @@ const run = async (chromy, options, test) => {
             try {
               logger.log(`Capturing screenshot of ${action.value} selector`);
               const png = await chromy.screenshotSelector(action.value);
-              await saveImage(test.name, png, test.type, options.screenshots, options.testReportPath);
+              saveImage(test.name, png, test.type, options.screenshots, options.testReportPath);
             } catch (error) {
               logger.error(error);
               return false;
@@ -62,8 +60,6 @@ const run = async (chromy, options, test) => {
         break;
       case actions.test:
         try {
-          logger.log(`comparing -> ${options.testReportPath}/${test.name}.png
-            and ${options.screenshots}/${test.name}.png`);
           const result = await compareImage(options, test.name);
           logger.log(result);
         } catch (error) {