+++ src/index.test.js
@@ -4,7 +4,7 @@ import Differencify from './index';
 import logger from './logger';
 
 let chromyCloseCallsCounter = 0;
-jest.mock('chromy', () => jest.fn().mockImplementation(() =>
+jest.mock('chromy', () => jest.fn(() =>
     ({
       goto: jest.fn(),
       close: jest.fn(() => { chromyCloseCallsCounter += 1; }),
@@ -78,7 +78,6 @@ describe('Differencify', () => {
     expect(loggerCalls[0]).toEqual('goto -> www.example.com');
     expect(loggerCalls[1]).toEqual('Capturing screenshot of whole DOM');
     expect(loggerCalls[2]).toEqual('screenshot saved in -> ./differencify_report/default.png');
-    expect(loggerCalls[4]).toEqual('Saving the diff image to disk');
     expect(writeFileSyncCalls).toEqual(['./differencify_report/default.png', 'png file']);
   });
   it('cleanup fn', async () => {